+++ src/main/java/leo/Leo.java
@@ -1,3 +1,9 @@
+package leo;
+
+import leo.commands.Command;
+import leo.storage.Storage;
+import leo.tasks.TaskList;
+
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
