+++ src/main/java/leo/LeoException.java
@@ -1,3 +1,5 @@
+package leo;
+
 public class LeoException extends Exception {
     public LeoException(String message) {
         super(message);
