+++ src/main/java/leo/commands/AddCommand.java
@@ -1,3 +1,12 @@
+package leo.commands;
+
+
+import leo.Ui;
+import leo.storage.Storage;
+import leo.tasks.Task;
+import leo.tasks.TaskList;
+import leo.LeoException;
+
 public class AddCommand extends Command {
     private final Task task;
