+++ src/main/java/leo/commands/Command.java
@@ -1,3 +1,10 @@
+package leo.commands;
+
+import leo.LeoException;
+import leo.Ui;
+import leo.storage.Storage;
+import leo.tasks.TaskList;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws LeoException;
