+++ src/main/java/leo/commands/DeleteCommand.java
@@ -1,3 +1,11 @@
+package leo.commands;
+
+import leo.Ui;
+import leo.storage.Storage;
+import leo.tasks.Task;
+import leo.tasks.TaskList;
+import leo.LeoException;
+
 public class DeleteCommand extends Command {
     private final int index;
