+++ src/main/java/leo/commands/UnmarkCommand.java
@@ -1,16 +1,23 @@
+package leo.commands;
+
+import leo.Ui;
+import leo.storage.Storage;
+import leo.tasks.TaskList;
+import leo.LeoException;
+
 import java.io.IOException;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
     private final int index;
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws LeoException {
         try {
-            tasks.get(index).markAsDone();
-            ui.showMarked(tasks.get(index));
+            tasks.get(index).markAsNotDone();
+            ui.showUnmarked(tasks.get(index));
             try {
                 storage.save(tasks.list());
             } catch (IOException e) {
