+++ src/main/java/leo/storage/Storage.java
@@ -1,3 +1,11 @@
+package leo.storage;
+
+import leo.tasks.Task;
+import leo.tasks.ToDo;
+import leo.tasks.Deadline;
+import leo.tasks.Event;
+import leo.LeoException;
+
 import java.io.IOException;
 import java.io.FileWriter;
 import java.nio.file.Files;
