+++ src/main/java/leo/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package leo.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
