+++ src/main/java/leo/tasks/Event.java
@@ -1,3 +1,5 @@
+package leo.tasks;
+
 public class Event extends Task {
     protected String from;
     protected String to;
