+++ src/main/java/leo/tasks/Task.java
@@ -1,3 +1,5 @@
+package leo.tasks;
+
 public class Task {
     protected String description;
     protected boolean isDone;
