+++ src/main/java/leo/tasks/ToDo.java
@@ -1,3 +1,5 @@
+package leo.tasks;
+
 public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
